requirements.txt
APScheduler==3.6.3
attrs==21.2.0
Automat==20.2.0
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
beautifulsoup4==4.9.3
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==8.0.1
colorama @ file:///tmp/build/80754af9/colorama_1607707115595/work
constantly==15.1.0
cryptography==3.4.7
cssselect==1.1.0
cycler==0.10.0
debugpy==1.3.0
decorator @ file:///tmp/build/80754af9/decorator_1621259047763/work
Flask==2.0.1
googlesearch-python==1.0.1
h2==3.2.0
hpack==3.0.0
hyperframe==5.2.0
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
ipykernel==6.0.1
ipython==7.25.0
ipython-genutils @ file:///tmp/build/80754af9/ipython_genutils_1606773439826/work
itemadapter==0.2.0
itemloaders==1.0.4
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
jmespath==0.10.0
joblib==1.0.1
jupyter-client @ file:///tmp/build/80754af9/jupyter_client_1616770841739/work
jupyter-core @ file:///C:/ci/jupyter_core_1612213515523/work
kiwisolver==1.3.1
lxml==4.6.3
MarkupSafe==2.0.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
nltk==3.6.2
numpy==1.21.0
pandas==1.3.0
parsel==1.6.0
parso==0.8.2
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==8.3.0
priority==1.3.0
prompt-toolkit==3.0.19
Protego==0.1.16
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
Pygments @ file:///tmp/build/80754af9/pygments_1621606182707/work
pyOpenSSL==20.0.1
pyparsing==2.4.7
python-dateutil @ file:///home/ktietz/src/ci/python-dateutil_1611928101742/work
python-telegram-bot==13.7
pytz==2021.1
pywin32==301
pyzmq==22.1.0
queuelib==1.6.1
regex==2021.7.6
requests==2.25.1
scikit-learn==0.24.2
scipy==1.7.0
Scrapy==2.5.0
service-identity==21.1.0
six @ file:///tmp/build/80754af9/six_1623709665295/work
soupsieve==2.2.1
threadpoolctl==2.1.0
tornado @ file:///C:/ci/tornado_1606924294691/work
tqdm==4.61.2
traitlets @ file:///home/ktietz/src/ci/traitlets_1611929699868/work
Twisted==21.2.0
twisted-iocpsupport==1.0.1
tzlocal==2.1
urllib3==1.26.6
w3lib==1.22.0
wcwidth @ file:///tmp/build/80754af9/wcwidth_1593447189090/work
Werkzeug==2.0.1
wincertstore==0.2
zope.interface==5.4.0
